+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - New: option to set an alternative base URL for the API calls
 - New: run diagnostics (network, accounts, settings, gem)
-- Fixed: --version if no user authorized
+- Fixed: situations when no user authorized
 - Fixed: --unauthorize if no valid username
 - Deprecated: Ayadn 1.x migration tools
 - Deprecated: "ayadn-app.net" domain name (use GitHub repository)